Sample: timer (Java UI with a Crystal backend)
This sample creates a Java activity that creates a View then runs a
Crystal application in its own thread, so it's separated from the
main thread and will never block. Updates to the UI is achieved via
messages passed to `android.os.Handler` that the Activity will
receive and process, since only the main thread is allowed to update
the UI.

This sample doesn't try to communicate actions or events from Java
to Crystal. This may require a Handler to pass messages, or require
`native` JNI functions —implemented in Crystal and running in the
main thread— that could pass messages through a thread-safe Queue,
received by a fiber —running in the crystal thread this time.